Switch from distutils to setuptools
Setuptools has a number of useful features and improvements over
distutils for packaging and installation. The first one being this
allows for a user to run `python setup.py develop --user` as an
alternative to pip install. This is often helpful for development as it
allows for linking with an egg to the user's local site-packages, which
enables updates to the package to take effect without reinstalling
during development.