Solve some introspection problems about the different type of words have the same name.
ex:
「select min from tbl1;」. Get introspection from 「select mi」.
There are two types of suggestion - function, and column have the same name. Here use reorders check priority to solve. And another case needs to resolve to function. ex:  「select min(min) from tbl1;」 get introspection from 「select mi」. When check it is a column or not. Look at the next token. If the next token is an instance of Parenthesis. And it is in the function list. Then would resolve it to function.